Multi-distro support
- adds in-memory embedded registry to serve registry image from "seed" config spec
- handles helm charts install/upgrade/rollback/uninstall via helm directly
- add tunnel capability to access zarf endpoints without ingress config
- flattens image collection to single tarball
- remove root / linux requirements for zarf deployments
- add K8s distro-detection and distro-specific seed image behaviors
- document local registry per https://github.com/kubernetes/enhancements/tree/master/keps/sig-cluster-lifecycle/generic/1755-communicating-a-local-registry
- reorganize logging, remove logrus, replace with pterm, add progress cli feedback
- add support for stacking helm values files
- normalize log levels: info, warn, debug, trace 
- update go mod deps
- enable full support for helm template -> argo or helm install directly
- move zarf state to a k8s secret in the zarf namespace
- support zarf destroy for 3rd-party clusters
- move crane image cache to sub folder under home
- add support for multiple users with htpasswd
- add path key for git-based helm charts
- remove need for PKI prompts on default init
- add storage class tracking in zarf state for multi-distro compatibility

+++ cli/cmd/destroy.go
@@ -5,7 +5,8 @@ import (
 	"os"
 	"regexp"
 
-	"github.com/defenseunicorns/zarf/cli/config"
+	"github.com/defenseunicorns/zarf/cli/internal/k8s"
+	"github.com/defenseunicorns/zarf/cli/internal/message"
 	"github.com/defenseunicorns/zarf/cli/internal/utils"
 
 	"github.com/spf13/cobra"
@@ -18,15 +19,26 @@ var destroyCmd = &cobra.Command{
 	Short: "Tear it all down, we'll miss you Zarf...",
 	Run: func(cmd *cobra.Command, args []string) {
 		burn()
-		_ = os.Remove(config.ZarfStatePath)
-		pattern := regexp.MustCompile(`(?mi)zarf-clean-.+\.sh$`)
-		scripts := utils.RecursiveFileList("/usr/local/bin", pattern)
-		// Iterate over al matching zarf-clean scripts and exec them
-		for _, script := range scripts {
-			// Run the matched script
-			_, _ = utils.ExecCommand(true, nil, script)
-			// Try to remove the script, but ignore any errors
-			_ = os.Remove(script)
+
+		state := k8s.LoadZarfState()
+		_ = os.Remove(".zarf-registry")
+
+		if state.ZarfAppliance {
+			// If Zarf deployed the cluster, burn it all down
+			pattern := regexp.MustCompile(`(?mi)zarf-clean-.+\.sh$`)
+			scripts := utils.RecursiveFileList("/usr/local/bin", pattern)
+			// Iterate over al matching zarf-clean scripts and exec them
+			for _, script := range scripts {
+				// Run the matched script
+				_, _ = utils.ExecCommand(true, nil, script)
+				// Try to remove the script, but ignore any errors
+				_ = os.Remove(script)
+			}
+		} else {
+			// If Zarf didn't deploy the cluster, only delete the ZarfNamespace
+			if err := k8s.DeleteZarfNamespace(); err != nil {
+				message.Fatal(err, "Unable to fully remove the zarf namespace from this cluster")
+			}
 		}
 		burn()
 	},
